can edit and delete doctors now

+++ main.py
@@ -59,8 +59,10 @@ def login_patient():
 @app.route('/admin_manage_doctors.html', methods=['POST'])
 def manage_doctor():
     form = request.form
-    if form["action"] == "delete":
+    if "action" in form.keys() and form["action"] == "delete":
         execute_query("delete from doctor where doc_id = " +form["docid"])
+    else:
+        execute_query('update doctor set first_name = "' + form['fname'] + '", last_name = "' + form['lname'] + '", aadhar_id = ' + form['adhar_id'] + ', chamber = "' + form['chamber'] + '", salary = ' + form['salary'] + ', timeslot = "' + form['timeslot'] + '" where doc_id = ' + form['docid'])
     return render_template("admin_manage_doctors.html", doctor_table=disp("select doc_id, first_name, last_name, aadhar_id, chamber, salary, dept_id, timeslot from doctor"))
 
 @app.route('/<page_type>', methods=['GET'])
@@ -71,6 +73,7 @@ def page(page_type):
     if(request.remote_addr in login.keys()):
         if(login[request.remote_addr] == 'a'):
             if(page_type == 'admin_manage_doctors.html'):
+                print(disp("select doc_id, first_name, last_name, aadhar_id, chamber, salary, dept_id, timeslot from doctor"))
                 return render_template(page_type, doctor_table=disp("select doc_id, first_name, last_name, aadhar_id, chamber, salary, dept_id, timeslot from doctor"))
             elif(page_type == 'admin_manage_patients.html'):
                 return render_template(page_type, patient_table=disp("select p_id, p_name, blood_grp, age, gender, ph_no, address, aadhar_id, p_history from patient"))
